depth prepass to render pass and mainshadowcasterpass commented to work properly on ios
also some comments on test project scripts to not fail

+++ com.unity.render-pipelines.lightweight/Shaders/Utils/ScreenSpaceShadows.shader
@@ -36,8 +36,8 @@ Shader "Hidden/Lightweight Render Pipeline/ScreenSpaceShadows"
 
         struct Varyings
         {
-            half4  positionCS   : SV_POSITION;
-            half4  uv           : TEXCOORD0;
+            float4  positionCS   : SV_POSITION;
+            float4  uv           : TEXCOORD0;
             UNITY_VERTEX_OUTPUT_STEREO
         };
